Update .travis.yml
Removed unnecessary update of apt-get repositories and download of cargo
(that's already included in rustup.sh). The last travis build failed
because the process ran for too long, so removing this cruft should shorten
it.

Finally, travis-ci doesn't have /usr/local/bin/ in the PATH. So, we use
the absolute path to cargo and screw PATH.

This is build is green on travis as of 7/25/2014 at 9:44pm PST.