Add missing context module
The context module was in `lib.rs`, but when the public API was modified
and cleaned up, it was removed from the library and never placed back
into the `lib.rs` or as a separate module file. I believe the intention
was to create a separate module file, i.e. `context.rs`.